util: Move RSS utilities into its own module
In preparation to add more utility functions.

- Tests and docstrings were adapted accordingly
- Preserve API by importing the RSS functions from lib/util .
